create mvc pattern start PE 03 refactor
