+++ package.json
@@ -3,6 +3,7 @@
   "version": "1.0.0",
   "description": "Portfolio project developed using HTML, CSS, and JavaScript, showcasing web design skills including layout, typography, and responsive design. Created as part of a web design course at CoderHouse with class ID 51270",
   "main": "index.js",
+  "type": "module",
   "scripts": {
     "sass": "sass --watch scss:css",
     "sassbuild": "sass --no-source-map --style=compressed scss:css",
